src/stock-advisor/requirements.txt
# Core
openai==1.85.0
    # via
    #   agents (pyproject.toml)
    #   autogen-ext
    #   langchain-openai
    #   openai-agents
    #   semantic-kernel
openai-agents==0.0.17
    # via agents (pyproject.toml)
openapi-core==0.19.5
    # via semantic-kernel
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.2
    # via openapi-core
python-dotenv>=1.0.1

# Web UI
gradio>=4.44.0

# Data & Computation
pandas>=2.2.2
numpy>=1.26.4
yfinance>=0.2.44
requests>=2.32.3

# Technical & Financial Analysis
ta>=0.11.0
scipy>=1.13.1

# Visualization
matplotlib>=3.9.2
plotly>=5.24.1

# NLP & Sentiment
transformers>=4.45.0
torch>=2.4.1
beautifulsoup4>=4.12.3

# Optional: Memory / Storage
chromadb>=0.5.5
sqlite-utils>=3.36.0

# Utilities
tqdm>=4.66.4
loguru>=0.7.2

# News APIs and Parsing
newspaper3k>=0.2.8
feedparser>=6.0.11

# Dev Tools
black>=24.8.0
flake8>=7.1.1
pytest>=7.4.3
textblob>=0.17.1